gitserver: Move conf.watch of factory and related code where its initialised
